refactor(frontend): 悬挂/挂起工具卡处理拆至 suspendedToolParts——messageParts 回到 1000 行守门内
